feat: remove Browser&Node handler options interface from types package
Because these interfaces are constructor interface for individual
http handler, they only need to be exposed from its own http handler
package
