+++ packages/types/src/http.ts
@@ -100,49 +100,3 @@ export interface ResolvedHttpResponse extends HttpResponse {
 export interface HttpHandlerOptions {
   abortSignal?: AbortSignal;
 }
-
-/**
- * Represents the http options that can be passed to a browser http client.
- */
-export interface BrowserHttpOptions {
-  /**
-   * The number of milliseconds a request can take before being automatically
-   * terminated.
-   */
-  requestTimeout?: number;
-}
-
-/**
- * Represents the http options that can be passed to a node http client.
- */
-export interface NodeHttpOptions {
-  /**
-   * The maximum time in milliseconds that the connection phase of a request
-   * may take before the connection attempt is abandoned.
-   */
-  connectionTimeout?: number;
-
-  /**
-   * The maximum time in milliseconds that a socket may remain idle before it
-   * is closed.
-   */
-  socketTimeout?: number;
-}
-
-/**
- * Represents the http2 options that can be passed to a node http2 client.
- */
-export interface NodeHttp2Options {
-  /**
-   * The maximum time in milliseconds that a stream may remain idle before it
-   * is closed.
-   */
-  requestTimeout?: number;
-
-  /**
-   * The maximum time in milliseconds that a session or socket may remain idle
-   * before it is closed.
-   * https://nodejs.org/docs/latest-v12.x/api/http2.html#http2_http2session_and_sockets
-   */
-  sessionTimeout?: number;
-}
